+++ examples/resources/morpheus_resource_pool_group/import.sh
@@ -1,0 +1,1 @@
+terraform import morpheus_resource_pool_group.tf_example_resource_pool_group 1